all apis added

+++ core/schemas/product.py
@@ -1,16 +1,17 @@
 from ninja import Schema
 from typing import Optional
+from decimal import Decimal
 
 class ProductCreateSchema(Schema):
     business_id: int
     name: str
     description: Optional[str] = None
-    price: float
+    price: Decimal
 
 class ProductUpdateSchema(Schema):
     name: Optional[str] = None
     description: Optional[str] = None
-    price: Optional[float] = None
+    price: Optional[Decimal]= None
     is_active: Optional[bool] = None
 
 class ProductOutSchema(Schema):
